route name changed and some basic stuff

+++ src/Designs/Draw/Functions/ResizeNode/ResizeNode.js
@@ -65,7 +65,7 @@ export function resize(element, node) {
             newnode.style.height = newWidth;
 
             buttons[4].style.top = el.clientY + "px";
-            buttons[4].style.left = el.clientX - parseInt(buttons[4].style.width) + "px";
+            buttons[4].style.left = el.clientX - parseFloat(buttons[4].style.width) + "px";
 
             buttons[5].style.top = el.clientY + "px";
             buttons[6].style.top = el.clientY + "px";
@@ -97,7 +97,7 @@ export function resize(element, node) {
             newnode.style.width = newWidth; 
             newnode.style.height = newHeight; 
 
-            buttons[4].style.top = el.clientY - parseInt(newnode.style.height) + "px";
+            buttons[4].style.top = el.clientY - parseFloat(newnode.style.height) + "px";
             buttons[6].style.left = el.clientX + "px";
             buttons[7].style.top = el.clientY + "px";
 
@@ -153,7 +153,7 @@ export function resize(element, node) {
             buttons[0].style.top = el.clientY + "px";
             buttons[1].style.top = el.clientY + "px";
 
-            const newHeight = parseInt(buttons[7].style.top) - parseInt(buttons[4].style.top) + "px";
+            const newHeight = parseFloat(buttons[7].style.top) - parseFloat(buttons[4].style.top) + "px";
             newnode.style.top = buttons[4].style.top;
 
             newnode.style.height = newHeight;
@@ -177,7 +177,7 @@ export function resize(element, node) {
             buttons[2].style.left = el.clientX - 8 + "px";
             buttons[6].style.left = el.clientX + "px";
 
-            const newWidth = parseInt(buttons[6].style.left) - parseInt(buttons[5].style.left);
+            const newWidth = parseFloat(buttons[6].style.left) - parseFloat(buttons[5].style.left);
             newnode.style.width = newWidth + "px";
             buttons[4].style.width = newWidth + "px";
             buttons[7].style.width = newWidth + "px";
@@ -196,7 +196,7 @@ export function resize(element, node) {
             buttons[2].style.top = el.clientY - 8 + "px";
             buttons[7].style.top = el.clientY + "px"
 
-            const newHeight = parseInt(buttons[7].style.top) - parseInt(buttons[4].style.top) + "px";
+            const newHeight = parseFloat(buttons[7].style.top) - parseFloat(buttons[4].style.top) + "px";
             newnode.style.height = newHeight;
 
             buttons[6].style.height = newHeight;
@@ -220,7 +220,7 @@ export function resize(element, node) {
             buttons[4].style.left = el.clientX + "px";
             buttons[7].style.left = el.clientX + "px";
 
-            const newWidth = parseInt(buttons[6].style.left) - parseInt(buttons[5].style.left) + "px";
+            const newWidth = parseFloat(buttons[6].style.left) - parseFloat(buttons[5].style.left) + "px";
             newnode.style.width = newWidth;
 
             buttons[4].style.width = newWidth;
